removed cpe database code
